Use RPIO package, fix no sensors.
Using RPIO package does not require running homebridge as root, which
is a really bad idea.

Also, fixing the logic for when no sensors are present.

Tested w/o sensors and running as user 'pi'. I added the logic to
initialize the INPUT ports too, but I did not test that.